requirements.txt
sounddevice
scipy
numpy
keyboard
openai-whisper
# ffmpeg-python
transformers 
accelerate 
datasets 
soundfile